+++ spec/lib/hyperion/formats_spec.rb
@@ -5,8 +5,14 @@ class Hyperion
     include Formats
 
     describe '#write' do
+      it 'returns the input if the input is a string' do
+        expect(write('hello', :json)).to eql 'hello'
+      end
       it 'writes json' do
         expect(write({'a' => 1}, :json)).to be_json_eql '{"a":1}'
+      end
+      it 'allows protobuf format but just passes it through' do
+        expect(write('x', :protobuf)).to eql 'x'
       end
     end
 
@@ -16,6 +22,9 @@ class Hyperion
       end
       it 'read json' do
         expect(read('{"a":1}', :json)).to eql({'a' => 1})
+      end
+      it 'allows protobuf format but just passes it throughn' do
+        expect(read('x', :protobuf)).to eql 'x'
       end
     end
   end