+++ CHANGES.md
@@ -4,3 +4,4 @@
 ### 0.0.17
 - If Hyperion.fake allow is passed a route, its block can now return an object instead of a rack-style response.
   The object is serialized according to the route's response descriptor.
+- Serialize the POST/PUT payload according to the route's payload descriptor.